requirements.txt
Flask==3.1.1
requests==2.32.4
python-dotenv==1.1.1
pdf2image==1.17.0
Pillow==11.2.1
reportlab==4.4.2
pypdf==5.6.1
psycopg2-binary==2.9.7
Flask-Login==0.6.3
Authlib==1.6.0
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
httplib2==0.22.0
